added middlewares and schemas separately
