+++ controllers/contacts.js
@@ -1,14 +1,6 @@
-const Joi = require("joi");
-
 const contacts = require("../models/contacts");
 
 const { HttpError, ctrlWrapper } = require("../helpers");
-
-const addSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().required(),
-  phone: Joi.string().required(),
-});
 
 const listContacts = async (req, res) => {
   const result = await contacts.listContacts();
@@ -25,10 +17,6 @@ const getContactById = async (req, res) => {
 };
 
 const addContact = async (req, res) => {
-  const { error } = addSchema.validate(req.body);
-  if (error) {
-    throw HttpError(400, "Missing required name field");
-  }
   const result = await contacts.addContact(req.body);
   res.status(201).json(result);
 };
