+++ routes/api/contacts.js
@@ -2,15 +2,17 @@ const express = require("express");
 
 const router = express.Router();
 const ctrl = require("../../controllers/contacts");
+const {validateBody} = require("../../middlewares");
+const schemas = require("../../schema/contacts");
 
 router.get("/", ctrl.listContacts);
 
 router.get("/:contactId", ctrl.getContactById);
 
-router.post("/", ctrl.addContact);
+router.post("/", validateBody(schemas.addSchema), ctrl.addContact);
+
+router.put("/:contactId", validateBody(schemas.addSchema), ctrl.updateContact);
 
 router.delete("/:contactId", ctrl.removeContact);
-
-router.put("/:contactId", ctrl.updateContact);
 
 module.exports = router;
